pyproject.toml
[project]
name = "saasflow-digital-fte"
version = "0.1.0"
description = "Incubation prototype for SaaSFlow AI Customer Success Agent"
authors = [
    {name = "AKE"}
]
dependencies = [
    "agents>=1.4.0",
    "aiokafka>=0.13.0",
    "aiosmtplib>=5.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "locust>=2.20.0",
    "mcp>=1.0.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.24.0",
    "openai-agents>=0.10.3",
    "pydantic[email]>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "production/tests"]
